Validering av arbeidsgiverperioder
