+++ api/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/api/innsending/InnsendingValidate.kt
@@ -1,14 +1,20 @@
 package no.nav.helsearbeidsgiver.inntektsmelding.api.innsending
 
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.FullLonnIArbeidsgiverPerioden
 import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.InnsendingRequest
 import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Inntekt
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Periode
 import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Refusjon
 import no.nav.helsearbeidsgiver.inntektsmelding.api.validation.isIdentitetsnummer
 import no.nav.helsearbeidsgiver.inntektsmelding.api.validation.isOrganisasjonsnummer
 import no.nav.helsearbeidsgiver.inntektsmelding.api.validation.isValidBehandlingsdager
+import org.valiktor.ConstraintViolationException
+import org.valiktor.functions.isEmpty
 import org.valiktor.functions.isGreaterThan
 import org.valiktor.functions.isGreaterThanOrEqualTo
 import org.valiktor.functions.isLessThan
+import org.valiktor.functions.isNotEmpty
 import org.valiktor.functions.isNotNull
 import org.valiktor.functions.isTrue
 import org.valiktor.functions.validate
@@ -22,14 +28,32 @@ fun InnsendingRequest.validate() {
         // Arbeidsgiver
         validate(InnsendingRequest::identitetsnummer).isNotNull()
         validate(InnsendingRequest::identitetsnummer).isIdentitetsnummer()
+        // Arbeidsgiverperioder
+        validate(InnsendingRequest::arbeidsgiverperioder).validateForEach {
+            validate(Periode::fom).isNotNull()
+            validate(Periode::tom).isNotNull()
+            validate(Periode::tom).isGreaterThanOrEqualTo(it.fom)
+        }
+        // Er tillatt å unngå arbeidsgiverperioder når:
+        // - arbeidsgiver ikke betaler lønn i arbeidsgiverperioden
+
+        println("Utbetaler: ${it.fullLønnIArbeidsgiverPerioden.utbetalerFullLønn}")
+        println("Perioder: ${it.arbeidsgiverperioder.size}")
+
+        if (it.fullLønnIArbeidsgiverPerioden.utbetalerFullLønn) {
+            println("utbetalerFullLønn")
+            validate(InnsendingRequest::arbeidsgiverperioder).isNotEmpty()
+        } else {
+            println("ikke fullLønn")
+        }
         // Fraværsperiode
         validate(InnsendingRequest::behandlingsdager).isValidBehandlingsdager() // Velg behandlingsdager
         // Egenmelding
         validate(InnsendingRequest::egenmeldingsperioder).validateForEach {
             it.fom
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Periode::fom).isNotNull()
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Periode::tom).isNotNull()
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Periode::tom).isGreaterThanOrEqualTo(it.fom)
+            validate(Periode::fom).isNotNull()
+            validate(Periode::tom).isNotNull()
+            validate(Periode::tom).isGreaterThanOrEqualTo(it.fom)
         }
         // Brutto inntekt
         validate(InnsendingRequest::inntekt).validate {
@@ -40,10 +64,16 @@ fun InnsendingRequest.validate() {
                 validate(Inntekt::endringÅrsak).isNotNull()
             }
         }
+        validate(InnsendingRequest::egenmeldingsperioder).validateForEach {
+            it.fom
+            validate(Periode::fom).isNotNull()
+            validate(Periode::tom).isNotNull()
+            validate(Periode::tom).isGreaterThanOrEqualTo(it.fom)
+        }
         // Betaler arbeidsgiver full lønn til arbeidstaker
         validate(InnsendingRequest::fullLønnIArbeidsgiverPerioden).validate {
             if (!it.utbetalerFullLønn) {
-                validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.FullLonnIArbeidsgiverPerioden::begrunnelse).isNotNull()
+                validate(FullLonnIArbeidsgiverPerioden::begrunnelse).isNotNull()
             }
         }
         // Betaler arbeidsgiver lønn under hele eller deler av sykefraværet
@@ -55,11 +85,11 @@ fun InnsendingRequest.validate() {
         }
         // Naturalytelser
         validate(InnsendingRequest::naturalytelser).validateForEach {
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse::naturalytelse).isNotNull()
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse::dato).isNotNull()
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse::beløp).isNotNull()
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse::beløp).isGreaterThan(0.0.toBigDecimal())
-            validate(no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse::beløp).isLessThan(1_000_000.0.toBigDecimal())
+            validate(Naturalytelse::naturalytelse).isNotNull()
+            validate(Naturalytelse::dato).isNotNull()
+            validate(Naturalytelse::beløp).isNotNull()
+            validate(Naturalytelse::beløp).isGreaterThan(0.0.toBigDecimal())
+            validate(Naturalytelse::beløp).isLessThan(1_000_000.0.toBigDecimal())
         }
         validate(InnsendingRequest::bekreftOpplysninger).isTrue()
     }
